Adicionando a função de likes
